2008-03-11  Tatsuhiro Tsujikawa  <tujikawa at rednoah dot com>
	Added summary log for peer announce storage.
	* src/DHTPeerAnnounceStorage.cc (handleTimeout)
	
	Removed cast
	* src/DefaultPeerStorage.cc (addPeer)

	Log exception message
	* src/DHTMessageTracker.cc (handleTimeout)

	Fixed casting
	* test/HttpRequestTest.cc
	* test/MetalinkProcessorTest.cc
	* test/PieceTest.cc
	* test/MetalinkEntryTest.cc
	* test/SequenceTest.cc

+++ src/DefaultPeerStorage.cc
@@ -81,7 +81,7 @@ bool DefaultPeerStorage::addPeer(const PeerHandle& peer) {
     logger->debug("Adding %s:%u is rejected because it has been already added.", peer->ipaddr.c_str(), peer->port);
     return false;
   }
-  if(peers.size() >= (size_t)maxPeerListSize) {
+  if(peers.size() >= maxPeerListSize) {
     deleteUnusedPeer(peers.size()-maxPeerListSize+1);
   }
   peers.push_front(peer);
